requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
langchain
langchain_groq
fastapi_mail
python-dotenv
jinja2